implemented some TextBox control stuff.
ImeModeChanged event was removed because it was used only to preserve
lyric textbox (which has some special show/hide logic over piano roll)
and it could be altered by TextChanged (not precise alternation but okay).

+++ src/Cadencii.Gui.Framework.Xwt/toolkit/VScrollBarImpl.cs
@@ -10,7 +10,7 @@ namespace Cadencii.Gui.Toolkit
 		}
 
 		int Cadencii.Gui.Toolkit.UiVScrollBar.LargeChange {
-			get { return this.PageIncrement; }
+			get { return (int) this.PageIncrement; }
 			set { this.PageIncrement = value; }
 		}
